requirements.txt
# GoQuant Trading Bot - Requirements
# This file lists all Python dependencies with their versions

python-telegram-bot==13.15
requests==2.31.0
websocket-client==1.6.1
aiohttp==3.8.5
asyncio-mqtt==0.11.0
asyncio==3.4.3
websockets==10.3